chore: resolve clippy warnings and suppress code quality alerts - Applied `cargo clippy --fix` to address idiom improvements and minor optimizations (e.g., useless_vec). - Added `allow` annotations for dead code, unused variables, and specific lint rules (ptr_arg, module_inception, etc.) to preserve the existing API and fallback data. - Verified the build is now warning-free with 0 errors.

+++ src/default/chile_rut.rs
@@ -5,7 +5,7 @@ use crate::config::FakerConfig;
 /// Generate a random Chilean RUT number
 pub fn rut() -> String {
     let config = FakerConfig::current();
-    let number = config.rand_range(1000000, 25000000) as u32;
+    let number = config.rand_range(1000000, 25000000);
     let dv = calculate_dv(number);
     format!("{}-{}", number, dv)
 }
